move MoleculeLJBase to its own file
